% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{compare_distributions}
\alias{compare_distributions}
\title{Compare Beta-Danish with Standard Distributions}
\usage{
compare_distributions(object)
}
\arguments{
\item{object}{A fitted `betadanish` object.}
}
\value{
A ranked data frame of model comparisons.
}
\description{
Fits standard lifetime distributions (Weibull, Log-Normal, Log-Logistic, Gamma, 
Exponential) using the `flexsurv` package and compares them to the Beta-Danish fit.
}
