% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{compare_models}
\alias{compare_models}
\title{Compare Nested Beta-Danish Models}
\usage{
compare_models(full_model, sub_model)
}
\arguments{
\item{full_model}{A fitted 4-parameter `betadanish` object.}

\item{sub_model}{A fitted 3-parameter `betadanish` object.}
}
\value{
A data frame with the test statistic and p-value.
}
\description{
Performs a Likelihood Ratio Test (LRT) between the 4-parameter full model 
and the 3-parameter submodel.
}
