% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_betadanish.R
\name{report_betadanish}
\alias{report_betadanish}
\title{Create a compact report from a BetaDanish model fit}
\usage{
report_betadanish(fit)
}
\arguments{
\item{fit}{A fitted BetaDanish model object.}
}
\value{
A list containing model summary information.
}
\description{
Create a compact report from a BetaDanish model fit
}
