% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_stats.R
\name{bd_order_stat_pdf}
\alias{bd_order_stat_pdf}
\title{Density of the r-th Order Statistic}
\usage{
bd_order_stat_pdf(x, r, n, a, b, c, k, log = FALSE)
}
\arguments{
\item{x}{Numeric vector of time points.}

\item{r}{Integer order (1 = minimum, n = maximum).}

\item{n}{Integer sample size.}

\item{a, b, c, k}{Positive parameters of the Beta-Danish distribution.}

\item{log}{Logical; if \code{TRUE} return the log-density.}
}
\value{
Numeric vector (or its log).
}
\description{
Evaluates the probability density function of the r-th order statistic
from a sample of size n drawn from the Beta-Danish distribution.
}
\examples{
tgrid <- seq(0.01, 5, length.out = 50)
bd_order_stat_pdf(tgrid, r = 5, n = 20,
                  a = 1.5, b = 2.5, c = 2, k = 1)

}
