% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{coef.betadanish}
\alias{coef.betadanish}
\title{Extract Coefficients}
\usage{
\method{coef}{betadanish}(object, ...)
}
\arguments{
\item{object}{An object of class `betadanish`.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A named numeric vector of maximum likelihood parameter estimates.
}
\description{
Extract Coefficients
}
