% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{gof_betadanish}
\alias{gof_betadanish}
\title{Goodness-of-Fit Statistics for Beta-Danish Models}
\usage{
gof_betadanish(object)
}
\arguments{
\item{object}{A fitted `betadanish` object.}
}
\value{
A list containing the Information Criteria and the K-S statistic.
}
\description{
Computes Information Criteria (AIC, BIC, HQIC, AICC) and the Kolmogorov-Smirnov 
(K-S) statistic for a fitted Beta-Danish model.
}
