% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{logLik.betadanish}
\alias{logLik.betadanish}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{betadanish}(object, ...)
}
\arguments{
\item{object}{An object of class `betadanish`.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
An object of class logLik containing the maximized log-likelihood value, with degrees of freedom and number of observations stored as attributes.
}
\description{
Extract Log-Likelihood
}
