% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{print.betadanish}
\alias{print.betadanish}
\title{Print Method for Beta-Danish Fit}
\usage{
\method{print}{betadanish}(x, ...)
}
\arguments{
\item{x}{An object of class `betadanish`.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print Method for Beta-Danish Fit
}
