% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{print.summary.betadanish}
\alias{print.summary.betadanish}
\title{Print Summary Method for Beta-Danish Fit}
\usage{
\method{print}{summary.betadanish}(x, ...)
}
\arguments{
\item{x}{An object of class `summary.betadanish`.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print Summary Method for Beta-Danish Fit
}
