% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{remission}
\alias{remission}
\title{Bladder Cancer Remission Times}
\format{
A data frame with 128 rows and 2 columns:
\describe{
  \item{time}{Remission time in months}
  \item{status}{Event indicator (1 = event occurred)}
}
}
\source{
Lee, E. T., & Wang, J. W. (2003). Statistical Methods for Survival Data Analysis (3rd ed.). Wiley.
}
\usage{
remission
}
\description{
Remission times (in months) for 128 bladder cancer patients. This is a
complete (uncensored) sample widely used in lifetime distribution literature
to demonstrate decreasing or right-skewed hazard rates.
}
\examples{
data(remission)
\donttest{
fit <- fit_betadanish(survival::Surv(time, status) ~ 1, data = remission)
summary(fit)
}
}
\keyword{datasets}
