% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary.betadanish}
\alias{summary.betadanish}
\title{Summary Method for Beta-Danish Fit}
\usage{
\method{summary}{betadanish}(object, ...)
}
\arguments{
\item{object}{An object of class `betadanish`.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summary Method for Beta-Danish Fit
}
