% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{vcov.betadanish}
\alias{vcov.betadanish}
\title{Extract Variance-Covariance Matrix}
\usage{
\method{vcov}{betadanish}(object, ...)
}
\arguments{
\item{object}{An object of class `betadanish`.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A numeric variance-covariance matrix for the estimated model parameters.
}
\description{
Extract Variance-Covariance Matrix
}
